pyproject.toml
[project]
name = "daily-info"
version = "0.1.0"
description = "Self-hosted research reading desk for papers, blogs, and posts."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "feedparser>=6.0.11",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "python-dateutil>=2.9.0",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.32",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

